backend/pyproject.toml
[project]
name = "fullrag-plus-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "google-genai>=1.70.0",
    "langchain-docling>=2.0.0",
    "langchain-opendataloader-pdf>=2.0.0",
    "liteparse>=1.2.1",
    "markdown-it-py>=3.0.0",
    "marker-pdf>=1.10.2",
    "markdownify>=0.13.0",
    "pydantic>=2.0.0",
    "python-docx>=1.0.0",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=3.0.0",
    "tiktoken>=0.7.0",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "pgvector>=0.3.6",
    "psycopg[binary]>=3.2.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.15.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "openai>=1.50.0",
    "python-json-logger>=3.0.0",
    "cachetools>=5.0",
    "apscheduler>=3.10.4",
    "streamlit>=1.35.0",
    "plotly>=5.0.0",
    "python-multipart>=0.0.26",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
pythonpath = ["."]
